+++ src/editor/dataset/constant/Placeholder.ts
@@ -5,5 +5,5 @@ export const defaultPlaceholderOption: Readonly<Required<IPlaceholder>> = {
   color: '#DCDFE6',
   opacity: 1,
   size: 16,
-  font: 'Yahei'
+  font: 'Microsoft YaHei'
 }